reto-10/atareao/requirements.txt
Pillow
pilgram
